Move Graph to unified file structure
