+++ src/app/pages/HomePage/Graph/Graph/graph-utils.ts
