+++ src/app/pages/HomePage/Graph/GraphTooltip/index.tsx
@@ -2,7 +2,7 @@ import { styled, useTheme } from '@mui/material/styles'
 import Tooltip, { TooltipProps, tooltipClasses } from '@mui/material/Tooltip'
 import { FC } from 'react'
 import Box from '@mui/material/Box'
-import { COLORS } from '../../../../styles/theme/colors'
+import { COLORS } from '../../../../../styles/theme/colors'
 import AccessTimeIcon from '@mui/icons-material/AccessTime'
 import Typography from '@mui/material/Typography'
 import AdjustIcon from '@mui/icons-material/Adjust'
@@ -11,9 +11,9 @@ import useMediaQuery from '@mui/material/useMediaQuery'
 import { useTranslation } from 'react-i18next'
 import { TFunction } from 'i18next'
 import * as React from 'react'
-import { RouteUtils } from '../../../utils/route-utils'
+import { RouteUtils } from '../../../../utils/route-utils'
 import { useNavigate } from 'react-router-dom'
-import { Layer } from '../../../../config'
+import { Layer } from '../../../../../config'
 
 export interface GraphTooltipStyledProps {
   isMobile: boolean
