+++ src/app/pages/HomePage/Graph/ParaTimeSelector/index.tsx
@@ -1,21 +1,21 @@
 import { FC, memo, useEffect, useRef, useState } from 'react'
 import { styled, useTheme } from '@mui/material/styles'
 import Box from '@mui/material/Box'
-import paratimeSelectorGlow from './images/paratime-selector-glow.svg'
-import paratimeSelectorGlobe from './images/paratime-selector-globe.svg'
-import { Graph } from './Graph'
+import paratimeSelectorGlow from '../images/paratime-selector-glow.svg'
+import paratimeSelectorGlobe from '../images/paratime-selector-globe.svg'
+import { Graph } from '../Graph'
 import Button from '@mui/material/Button'
 import { useTranslation } from 'react-i18next'
-import { ParaTimeSelectorStep } from './types'
-import { ParaTimeSelectorUtils } from './para-time-selector-utils'
+import { ParaTimeSelectorStep } from '../types'
+import { ParaTimeSelectorUtils } from '../para-time-selector-utils'
 import Fade from '@mui/material/Fade'
 import useMediaQuery from '@mui/material/useMediaQuery'
 import QuickPinchZoom, { make3dTransformValue, UpdateAction } from 'react-quick-pinch-zoom'
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft'
-import { GraphUtils } from './Graph/graph-utils'
+import { GraphUtils } from '../Graph/graph-utils'
 import useResizeObserver from 'use-resize-observer'
-import HelpScreen from './HelpScreen'
-import { Layer } from '../../../../config'
+import HelpScreen from '../HelpScreen'
+import { Layer } from '../../../../../config'
 import { GraphTooltipMobile } from '../GraphTooltip/GraphTooltipMobile'
 
 interface ParaTimeSelectorBaseProps {
