+++ src/app/pages/HomePage/Graph/para-time-selector-utils.ts
