+++ src/app/pages/HomePage/Graph/types.ts
